smartspend/backend/requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.6
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
packaging==24.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
redis==5.0.4
sqlparse==0.5.0
whitenoise==6.6.0
